pyproject.toml
[project]
name = "math-compact"
version = "0.1.0"
description = "Minimal synthetic math dataset generator and compact transformer training"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "datasets>=2.10.0",
    "accelerate>=0.20.0",
    "einops>=0.6.0",
    "omegaconf>=2.2.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
    "scikit-learn>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "wandb>=0.15.0",
]
